Add isort script
isort is a little utility that automatically sort imports in python
files. It comes to complement yapf on having a consistent coding style.
run_isort.sh is essentially the same script as run_yapf except it runs
isort instead of yapf.
isort is also now in the requirements.txt file.