Add drawing of rotation buttons and interactions
Bring the sketching on the UI the rotation buttons, and the possiblity
for the user to click on them to rotate the disks.
This is done by firstly pre-generating all the data that we might need
to draw and interact with the buttons, then, when in the drawing loop,
we exploit this data to draw the buttons and call the needed routines
when clicking on them.

Also explicit the comments of some existing subroutines.