Rehandle use of globals to limit their number
Globals were becoming invasive, illogical and unreadable. In order to
put down this issue, I heavely reduced the number of global variables.
It results in more heavy computation when drawing, thus a way slower
interface.

This new issue can be solved by moving the computation of dimensions and
positions done when drawing to a data structure generated at setup. The
most of the cpu time may be used when drawing the cylinder, thus we
should move it a data structure of this kind.